Fix menu behavior
A bug in headless UI @ 1.7.7 (latest) did not close the menu after clicking a link. Because "Link" from Inertia has an event.preventDefault(), it does not allow for the event to bubble to the headless UI menu so that it can close. This is supposed to be solved by exposing a "close" function on the headless ui menu to manually do this, but a recent update has caused a regression so that the close function is no longer working. It is currently being tracked on their github under issue #2243.